Add tailwind and create prisma schema
