--- .env.example
@@ -1,3 +1,0 @@
-# DATABASE_URL=file:./dev.db
-# TEST_DATABASE_URL=file:./.redwood/test.db
-# PRISMA_HIDE_UPDATE_MESSAGE=true
