+++ .prettierrc
@@ -1,5 +1,6 @@
 {
   "endOfLine": "lf",
   "singleQuote": false,
-  "tabWidth": 2
+  "tabWidth": 2,
+  "trailingComma": "es5"
 }
