+++ web/config/postcss.config.js
@@ -1,0 +1,8 @@
+const path = require('path')
+
+module.exports = {
+  plugins: [
+    require('tailwindcss')(path.resolve(__dirname, '../tailwind.config.js')),
+    require('autoprefixer'),
+  ],
+}
