+++ web/src/index.css
@@ -1,0 +1,13 @@
+/**
+ * START --- TAILWIND GENERATOR EDIT
+ *
+ * `yarn rw setup tailwind` placed these imports here
+ * to inject Tailwind's styles into your CSS.
+ * For more information, see: https://tailwindcss.com/docs/installation#add-tailwind-to-your-css
+ */
+@import "tailwindcss/base";
+@import "tailwindcss/components";
+@import "tailwindcss/utilities";
+/**
+ * END --- TAILWIND GENERATOR EDIT
+ */
