+++ web/src/pages/FatalErrorPage/FatalErrorPage.tsx
